refactor!: move stuff away from the root package
