+++ alchemist-euclidean-geometry/src/main/kotlin/it/unibo/alchemist/model/implementations/actions/HeadTowardTarget.kt
@@ -9,12 +9,12 @@
 
 package it.unibo.alchemist.model.implementations.actions
 
-import it.unibo.alchemist.toPosition
 import it.unibo.alchemist.model.interfaces.Context
 import it.unibo.alchemist.model.interfaces.Molecule
 import it.unibo.alchemist.model.interfaces.Node
 import it.unibo.alchemist.model.interfaces.Reaction
 import it.unibo.alchemist.model.interfaces.environments.Physics2DEnvironment
+import it.unibo.alchemist.model.util.AnyExtension.toPosition
 import org.apache.commons.math3.util.FastMath.toRadians
 import kotlin.math.abs
 import kotlin.math.cos
