+++ alchemist-euclidean-geometry/src/main/kotlin/it/unibo/alchemist/model/implementations/movestrategies/RandomTarget.kt
@@ -10,11 +10,11 @@
 package it.unibo.alchemist.model.implementations.movestrategies
 
 import it.unibo.alchemist.model.implementations.positions.Euclidean2DPosition
-import it.unibo.alchemist.nextDouble
 import it.unibo.alchemist.model.interfaces.Environment
 import it.unibo.alchemist.model.interfaces.Node
 import it.unibo.alchemist.model.interfaces.Reaction
 import it.unibo.alchemist.model.interfaces.environments.Environment2DWithObstacles
+import it.unibo.alchemist.model.util.RandomGeneratorExtension.nextDouble
 import kotlin.math.cos
 import kotlin.math.sin
 import org.apache.commons.math3.distribution.RealDistribution
