+++ alchemist-euclidean-geometry/src/main/kotlin/it/unibo/alchemist/model/interfaces/geometry/euclidean2d/Segment2D.kt
@@ -1,8 +1,8 @@
 package it.unibo.alchemist.model.interfaces.geometry.euclidean2d
 
-import it.unibo.alchemist.fuzzyIn
+import it.unibo.alchemist.model.util.DoubleExtension.fuzzyIn
+import it.unibo.alchemist.model.util.RangeExtension.rangeFromUnordered
 import it.unibo.alchemist.model.implementations.geometry.euclidean2d.SlopeInterceptLine2D
-import it.unibo.alchemist.rangeFromUnordered
 import it.unibo.alchemist.model.interfaces.geometry.Vector2D
 import org.danilopianini.lang.MathUtils.fuzzyEquals
 import java.lang.UnsupportedOperationException
