+++ alchemist-implementationbase/src/main/java/it/unibo/alchemist/model/api/SupportedIncarnations.java
@@ -1,12 +1,14 @@
 /*
- * Copyright (C) 2010-2019, Danilo Pianini and contributors listed in the main project's alchemist/build.gradle file.
+ * Copyright (C) 2010-2022, Danilo Pianini and contributors
+ * listed, for each module, in the respective subproject's build.gradle.kts file.
  *
  * This file is part of Alchemist, and is distributed under the terms of the
  * GNU General Public License, with a linking exception,
  * as described in the file LICENSE in the Alchemist distribution's top directory.
  */
-package it.unibo.alchemist;
+package it.unibo.alchemist.model.api;
 
+import it.unibo.alchemist.util.ClassPathScanner;
 import it.unibo.alchemist.model.interfaces.Incarnation;
 import it.unibo.alchemist.model.interfaces.Position;
 import org.jooq.lambda.Unchecked;
