+++ alchemist-implementationbase/src/main/java/it/unibo/alchemist/model/implementations/environments/AbstractEnvironment.java
@@ -14,8 +14,8 @@ import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import gnu.trove.map.hash.TIntObjectHashMap;
 import gnu.trove.set.TIntSet;
 import gnu.trove.set.hash.TIntHashSet;
-import it.unibo.alchemist.SupportedIncarnations;
 import it.unibo.alchemist.core.interfaces.Simulation;
+import it.unibo.alchemist.model.api.SupportedIncarnations;
 import it.unibo.alchemist.model.interfaces.Environment;
 import it.unibo.alchemist.model.interfaces.Incarnation;
 import it.unibo.alchemist.model.interfaces.Layer;
