+++ alchemist-implementationbase/src/main/java/it/unibo/alchemist/model/implementations/linkingrules/ClosestN.java
@@ -12,7 +12,7 @@ import com.google.common.cache.CacheBuilder;
 import com.google.common.collect.MinMaxPriorityQueue;
 import com.google.common.primitives.Doubles;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import it.unibo.alchemist.BugReporting;
+import it.unibo.alchemist.util.BugReporting;
 import it.unibo.alchemist.model.implementations.neighborhoods.Neighborhoods;
 import it.unibo.alchemist.model.interfaces.Environment;
 import it.unibo.alchemist.model.interfaces.LinkingRule;
