+++ alchemist-implementationbase/src/main/kotlin/it/unibo/alchemist/util/BugReporting.kt
@@ -1,12 +1,13 @@
 /*
- * Copyright (C) 2010-2021, Danilo Pianini and contributors
- * listed in the main project's alchemist/build.gradle.kts file.
+ * Copyright (C) 2010-2022, Danilo Pianini and contributors
+ * listed, for each module, in the respective subproject's build.gradle.kts file.
  *
  * This file is part of Alchemist, and is distributed under the terms of the
  * GNU General Public License, with a linking exception,
  * as described in the file LICENSE in the Alchemist distribution's top directory.
  */
-package it.unibo.alchemist
+
+package it.unibo.alchemist.util
 
 /**
  * Entrypoint for printing meaningful debug information in case of failed internal consistency checks.
