+++ alchemist-loading/src/main/kotlin/it/unibo/alchemist/loader/LoadAlchemist.kt
@@ -9,7 +9,7 @@
 
 package it.unibo.alchemist.loader
 
-import it.unibo.alchemist.ClassPathScanner
+import it.unibo.alchemist.util.ClassPathScanner
 import it.unibo.alchemist.loader.m2m.SimulationModel
 import java.io.File
 import java.io.InputStream
