+++ alchemist-loading/src/test/java/it/unibo/alchemist/test/TestYAMLLoader.java
@@ -7,7 +7,7 @@
  */
 package it.unibo.alchemist.test;
 
-import it.unibo.alchemist.ClassPathScanner;
+import it.unibo.alchemist.util.ClassPathScanner;
 import it.unibo.alchemist.SupportedIncarnations;
 import it.unibo.alchemist.core.implementations.Engine;
 import it.unibo.alchemist.core.interfaces.Simulation;
