+++ alchemist-swingui/src/main/java/it/unibo/alchemist/boundary/gui/effects/EffectBuilder.java
@@ -8,7 +8,7 @@
 package it.unibo.alchemist.boundary.gui.effects;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import it.unibo.alchemist.ClassPathScanner;
+import it.unibo.alchemist.util.ClassPathScanner;
 import it.unibo.alchemist.boundary.gui.AlchemistSwingUI;
 import it.unibo.alchemist.boundary.l10n.LocalizedResourceBundle;
 import org.slf4j.Logger;
