+++ javadoc-io.json
@@ -3,6 +3,10 @@
     "first": "https://javadoc.io/doc/com.github.ben-manes.caffeine/caffeine/3.0.5",
     "second": "https://javadoc.io/doc/com.github.ben-manes.caffeine/caffeine/3.0.5/element-list"
   },
+  "com.github.ben-manes.caffeine/caffeine/3.0.6": {
+    "first": "https://javadoc.io/doc/com.github.ben-manes.caffeine/caffeine/3.0.6",
+    "second": "https://javadoc.io/doc/com.github.ben-manes.caffeine/caffeine/3.0.6/element-list"
+  },
   "com.github.cb372/scalacache-core_2.13/0.28.0": {
     "first": "https://javadoc.io/doc/com.github.cb372/scalacache-core_2.13/0.28.0"
   },
@@ -12,6 +16,10 @@
   "com.github.davidmoten/rtree/0.8.7": {
     "first": "https://javadoc.io/doc/com.github.davidmoten/rtree/0.8.7",
     "second": "https://javadoc.io/doc/com.github.davidmoten/rtree/0.8.7/package-list"
+  },
+  "com.github.davidmoten/rtree/0.9": {
+    "first": "https://javadoc.io/doc/com.github.davidmoten/rtree/0.9",
+    "second": "https://javadoc.io/doc/com.github.davidmoten/rtree/0.9/element-list"
   },
   "com.github.jiconfont/jiconfont-google_material_design_icons/2.2.0.2": {
     "first": "https://javadoc.io/doc/com.github.jiconfont/jiconfont-google_material_design_icons/2.2.0.2",
@@ -25,13 +33,25 @@
     "first": "https://javadoc.io/doc/com.google.code.gson/gson/2.8.9",
     "second": "https://javadoc.io/doc/com.google.code.gson/gson/2.8.9/element-list"
   },
+  "com.google.code.gson/gson/2.9.0": {
+    "first": "https://javadoc.io/doc/com.google.code.gson/gson/2.9.0",
+    "second": "https://javadoc.io/doc/com.google.code.gson/gson/2.9.0/element-list"
+  },
   "com.google.guava/guava/31.0.1-jre": {
     "first": "https://javadoc.io/doc/com.google.guava/guava/31.0.1-jre",
     "second": "https://javadoc.io/doc/com.google.guava/guava/31.0.1-jre/element-list"
   },
+  "com.google.guava/guava/31.1-jre": {
+    "first": "https://javadoc.io/doc/com.google.guava/guava/31.1-jre",
+    "second": "https://javadoc.io/doc/com.google.guava/guava/31.1-jre/element-list"
+  },
   "com.graphhopper/graphhopper-core/4.0": {
     "first": "https://javadoc.io/doc/com.graphhopper/graphhopper-core/4.0",
     "second": "https://javadoc.io/doc/com.graphhopper/graphhopper-core/4.0/package-list"
+  },
+  "com.graphhopper/graphhopper-core/5.0": {
+    "first": "https://javadoc.io/doc/com.graphhopper/graphhopper-core/5.0",
+    "second": "https://javadoc.io/doc/com.graphhopper/graphhopper-core/5.0/package-list"
   },
   "com.javadocmd/simplelatlng/1.3.1": {
     "first": "https://javadoc.io/doc/com.javadocmd/simplelatlng/1.3.1",
@@ -70,6 +90,10 @@
     "first": "https://javadoc.io/doc/io.github.classgraph/classgraph/4.8.138",
     "second": "https://javadoc.io/doc/io.github.classgraph/classgraph/4.8.138/package-list"
   },
+  "io.github.classgraph/classgraph/4.8.143": {
+    "first": "https://javadoc.io/doc/io.github.classgraph/classgraph/4.8.143",
+    "second": "https://javadoc.io/doc/io.github.classgraph/classgraph/4.8.143/package-list"
+  },
   "io.jenetics/jpx/2.3.0": {
     "first": "https://javadoc.io/doc/io.jenetics/jpx/2.3.0",
     "second": "https://javadoc.io/doc/io.jenetics/jpx/2.3.0/element-list"
@@ -77,6 +101,10 @@
   "it.unimi.dsi/dsiutils/2.6.17": {
     "first": "https://javadoc.io/doc/it.unimi.dsi/dsiutils/2.6.17",
     "second": "https://javadoc.io/doc/it.unimi.dsi/dsiutils/2.6.17/element-list"
+  },
+  "it.unimi.dsi/dsiutils/2.7.0": {
+    "first": "https://javadoc.io/doc/it.unimi.dsi/dsiutils/2.7.0",
+    "second": "https://javadoc.io/doc/it.unimi.dsi/dsiutils/2.7.0/element-list"
   },
   "net.anwiba.commons.swing.icons/org.oxygen.oxygen-icons/4.13.0-1.1.158": {
     "first": "https://javadoc.io/doc/net.anwiba.commons.swing.icons/org.oxygen.oxygen-icons/4.13.0-1.1.158"
@@ -133,6 +161,10 @@
     "first": "https://javadoc.io/doc/org.danilopianini/jirf/0.4.5",
     "second": "https://javadoc.io/doc/org.danilopianini/jirf/0.4.5/element-list"
   },
+  "org.danilopianini/jirf/0.4.6": {
+    "first": "https://javadoc.io/doc/org.danilopianini/jirf/0.4.6",
+    "second": "https://javadoc.io/doc/org.danilopianini/jirf/0.4.6/element-list"
+  },
   "org.danilopianini/thread-inheritable-resource-loader/0.3.5": {
     "first": "https://javadoc.io/doc/org.danilopianini/thread-inheritable-resource-loader/0.3.5",
     "second": "https://javadoc.io/doc/org.danilopianini/thread-inheritable-resource-loader/0.3.5/package-list"
@@ -172,6 +204,10 @@
     "first": "https://javadoc.io/doc/org.mongodb/mongodb-driver-sync/4.4.1",
     "second": "https://javadoc.io/doc/org.mongodb/mongodb-driver-sync/4.4.1/element-list"
   },
+  "org.mongodb/mongodb-driver-sync/4.5.1": {
+    "first": "https://javadoc.io/doc/org.mongodb/mongodb-driver-sync/4.5.1",
+    "second": "https://javadoc.io/doc/org.mongodb/mongodb-driver-sync/4.5.1/element-list"
+  },
   "org.scala-lang/scala-compiler/2.13.8": {
     "first": "https://javadoc.io/doc/org.scala-lang/scala-compiler/2.13.8"
   },
@@ -185,6 +221,10 @@
   "org.slf4j/slf4j-api/1.7.35": {
     "first": "https://javadoc.io/doc/org.slf4j/slf4j-api/1.7.35",
     "second": "https://javadoc.io/doc/org.slf4j/slf4j-api/1.7.35/package-list"
+  },
+  "org.slf4j/slf4j-api/1.7.36": {
+    "first": "https://javadoc.io/doc/org.slf4j/slf4j-api/1.7.36",
+    "second": "https://javadoc.io/doc/org.slf4j/slf4j-api/1.7.36/package-list"
   },
   "org.yaml/snakeyaml/1.30": {
     "first": "https://javadoc.io/doc/org.yaml/snakeyaml/1.30",
