+++ src/test/kotlin/TestWebsiteCodeSnippets.kt
@@ -5,7 +5,7 @@ import io.kotest.matchers.collections.shouldNotBeEmpty
 import io.kotest.matchers.nulls.beNull
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldNot
-import it.unibo.alchemist.ClassPathScanner
+import it.unibo.alchemist.util.ClassPathScanner
 import it.unibo.alchemist.core.implementations.Engine
 import it.unibo.alchemist.loader.LoadAlchemist
